something wrong with do_intersect method + otherwise working fine

+++ glwidget.h
@@ -12,8 +12,8 @@ public:
     void paintGL();
     void resizeGL(int w, int h);
 
-    double transformX(int x, int width);
-    double transformY(int y, int height);
+    double transformX(double x, double width);
+    double transformY(double y, double height);
 };
 
 #endif // GLWIDGET_H
